fix(libltfs): report the incremental-to-full fallback where the type is resolved, reset the count at mount
Review findings on the full_index_interval change:
- ALB0286W was unreachable: ltfs_sync_index() resolves the index type
  before ltfs_write_index() sees it. The warning is now issued by
  ltfs_sync_index(), only when an index is actually written.
- full_index_to_go is reset when ltfs_mount() has read the index, so a
  cartridge mounted on a revalidation starts its own count.
- The read of full_index_to_go under the read lock is documented (two
  overlapping syncs may both write a full index, never a wrong one).
- The altfs option field is int64_t like the library API.
- The invalid-option test no longer formats a tape per value.

+++ src/cmd/altfs/altfs.c
@@ -1163,7 +1163,7 @@ int single_drive_main(struct fuse_args *args, struct ltfs_fuse_data *priv)
 		/* Index type policy of the periodic sync and the sync on close */
 		ltfs_set_full_index_interval(priv->full_index_interval, priv->data);
 		if (priv->full_index_interval_str)
-			ltfsmsg(AFS0149I, priv->full_index_interval);
+			ltfsmsg(AFS0149I, (long long)priv->full_index_interval);
 
 		/* Validate symbolic link type */
 		priv->data->livelink = false;
